+++ tomcat/src/main/java/com/techcourse/controller/RegisterController.java
@@ -1,13 +1,16 @@
-package org.apache.catalina.controller;
+package com.techcourse.controller;
 
 import com.techcourse.db.InMemoryUserRepository;
 import com.techcourse.model.Account;
+import com.techcourse.model.Email;
 import com.techcourse.model.Password;
+import com.techcourse.model.User;
 import java.io.IOException;
 import java.net.URISyntaxException;
 import java.nio.file.Files;
 import java.nio.file.Paths;
 import java.util.Map;
+import org.apache.catalina.controller.AbstractController;
 import org.apache.catalina.resolver.StaticResourceResolver;
 import org.apache.catalina.resolver.StaticResourceResolver.ResolvedResource;
 import org.apache.coyote.http11.request.Http11Request;
@@ -17,10 +20,10 @@ import org.apache.coyote.http11.session.SessionManager;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-public class LoginController extends AbstractController {
+public class RegisterController extends AbstractController {
 
     private static final String COOKIE = "JSESSIONID=";
-    private static final Logger log = LoggerFactory.getLogger(LoginController.class);
+    private static final Logger log = LoggerFactory.getLogger(RegisterController.class);
 
     private final StaticResourceResolver resolver = new StaticResourceResolver();
 
@@ -38,7 +41,7 @@ public class LoginController extends AbstractController {
 
             return Http11Response.ok("text/html;charset=utf-8", body);
         } catch (IOException | URISyntaxException e) {
-            log.error("로그인 페이지 읽기 실패", e);
+            log.error("회원가입 페이지 읽기 실패", e);
             return Http11Response.serverError();
         }
     }
@@ -48,19 +51,19 @@ public class LoginController extends AbstractController {
         final Map<String, String> params = request.extractRequestBodyParams();
         final Account account = new Account(params.get("account"));
         final Password password = new Password(params.get("password"));
+        final Email email = new Email(params.get("email"));
+        final User userEntity = new User(account, password, email);
+        InMemoryUserRepository.save(userEntity);
 
-        final var loginSuccess = InMemoryUserRepository.findByAccount(account)
-                .filter(user -> user.checkPassword(password));
+        boolean registerSuccess = InMemoryUserRepository.findByAccount(account)
+                .filter(user -> user.checkAccount(account) && user.checkPassword(password))
+                .isPresent();
 
-        if (loginSuccess.isPresent()) {
-            log.info("로그인 성공 - {}", account);
-            SessionManager.invalidateAllForUser(account);
-            Session newSession = SessionManager.create();
-            newSession.setAttribute("user", loginSuccess.get());
-
-            return Http11Response.redirect("/index.html", COOKIE + newSession.getId());
+        if (registerSuccess) {
+            log.info("회원가입 성공 - {}", userEntity);
+            return Http11Response.redirect("/login.html");
         }
-        log.warn("로그인 실패 - {}", account);
+        log.info("회원가입 실패  - {}", userEntity);
         return Http11Response.redirect("/401.html");
     }
 }
