+++ tomcat/src/main/java/com/techcourse/controller/StaticResourcesController.java
@@ -1,9 +1,10 @@
-package org.apache.catalina.controller;
+package com.techcourse.controller;
 
 import java.io.IOException;
 import java.net.URISyntaxException;
 import java.nio.file.Files;
 import java.nio.file.Paths;
+import org.apache.catalina.controller.AbstractController;
 import org.apache.catalina.resolver.StaticResourceResolver;
 import org.apache.catalina.resolver.StaticResourceResolver.ResolvedResource;
 import org.apache.coyote.http11.request.Http11Request;
